pyproject.toml
[tool.poetry]
name = "repl_nix_tgdex"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.8.3"
aiohttp-jinja2 = "^1.5"
aiohttp-session = "^2.11.0"
Jinja2 = "^3.1.2"
python-dotenv = "^0.21.0"
Telethon = "^1.25.2"
wget = "^3.2"
Pillow = "^9.2.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
